+++ snuba/cli/replacer.py
@@ -39,7 +39,7 @@ def replacer(replacements_topic, consumer_group, bootstrap_server, clickhouse_ho
 
     import sentry_sdk
     from snuba import util
-    from snuba.clickhouse import ClickhousePool
+    from snuba.clickhouse.native import ClickhousePool
     from batching_kafka_consumer import BatchingKafkaConsumer
     from snuba.replacer import ReplacerWorker
