+++ snuba/clickhouse/__init__.py
@@ -1,0 +1,1 @@
+DATETIME_FORMAT = "%Y-%m-%d %H:%M:%S"
