+++ snuba/datasets/outcomes.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 import uuid
 
-from snuba.clickhouse import (
+from snuba.clickhouse.columns import (
     ColumnSet,
     DateTime,
     LowCardinality,
