+++ snuba/perf.py
@@ -59,7 +59,7 @@ def get_messages(events_file):
 def run(events_file, dataset, repeat=1,
         profile_process=False, profile_write=False):
     from snuba.consumer import ConsumerWorker
-    from snuba.clickhouse import ClickhousePool
+    from snuba.clickhouse.native import ClickhousePool
 
     ClickhousePool().execute(dataset.get_schema().get_local_table_definition())
