+++ snuba/util.py
@@ -15,7 +15,6 @@ import _strptime  # NOQA fixes _strptime deferred import issue
 import time
 
 from snuba import clickhouse, schemas, settings, state
-from snuba.clickhouse import escape_col
 from snuba.reader import NativeDriverReader
 
 
@@ -28,6 +27,9 @@ QUOTED_LITERAL_RE = re.compile(r"^'.*'$")
 ESCAPE_STRING_RE = re.compile(r"(['\\])")
 SAFE_FUNCTION_RE = re.compile(r'-?[a-zA-Z_][a-zA-Z0-9_]*$')
 TOPK_FUNCTION_RE = re.compile(r'^top([1-9]\d*)$')
+ESCAPE_COL_RE = re.compile(r"([`\\])")
+NEGATE_RE = re.compile(r'^(-?)(.*)$')
+SAFE_COL_RE = re.compile(r'^-?[a-zA-Z_][a-zA-Z0-9_\.]*$')
 
 
 class InvalidConditionException(Exception):
@@ -46,6 +48,21 @@ def string_col(dataset, col):
         return escape_col(col)
     else:
         return 'toString({})'.format(escape_col(col))
+
+
+def escape_col(col):
+    if not col:
+        return col
+    elif SAFE_COL_RE.match(col):
+        # Column is safe to use without wrapping.
+        return col
+    else:
+        # Column needs special characters escaped, and to be wrapped with
+        # backticks. If the column starts with a '-', keep that outside the
+        # backticks as it is not part of the column name, but used by the query
+        # generator to signify the sort order if we are sorting by this column.
+        col = ESCAPE_COL_RE.sub(r"\\\1", col)
+        return u'{}`{}`'.format(*NEGATE_RE.match(col).groups())
 
 
 def parse_datetime(value, alignment=1):
@@ -265,6 +282,8 @@ def conditions_expr(dataset, conditions, body, depth=0):
     Expansion of columns is handled, as is replacement of columns with aliases,
     if the column has already been expanded and aliased elsewhere.
     """
+    from snuba.clickhouse.columns import Array
+
     if not conditions:
         return ''
 
@@ -299,7 +318,7 @@ def conditions_expr(dataset, conditions, body, depth=0):
         if (
             isinstance(lhs, str) and
             lhs in columns and
-            isinstance(columns[lhs].type, clickhouse.Array) and
+            isinstance(columns[lhs].type, Array) and
             columns[lhs].base_name != body.get('arrayjoin') and
             not isinstance(lit, (list, tuple))
         ):
