+++ tests/base.py
@@ -8,7 +8,7 @@ from confluent_kafka import TopicPartition
 
 from snuba import settings
 from snuba.datasets.factory import get_dataset
-from snuba.clickhouse import ClickhousePool
+from snuba.clickhouse.native import ClickhousePool
 from snuba.redis import redis_client
 from snuba.perf import FakeKafkaMessage
